Refactor inventory (#1251)
* As noted in comments from previous PR, just some reorg kept separate from the functionality PR

* Correct matching
