+++ imports/plugins/included/inventory/server/hooks/hooks.js
