+++ imports/plugins/included/inventory/server/methods/inventory.js
@@ -1,287 +1,169 @@
 import { Meteor } from "meteor/meteor";
+import { check, Match } from "meteor/check";
+import { Catalog } from "/lib/api";
 import { Inventory } from "/lib/collections";
 import * as Schemas from "/lib/collections/schemas";
 import { Logger, Reaction } from "/server/api";
 
-// Disabled for now, needs more testing.
 
-// // Define a rate limiting rule that matches update attempts by non-admin users
-// const addReserveRule = {
-//   userId: function (userId) {
-//     return Roles.userIsInRole(userId, "createProduct", Reaction.getShopId());
-//   },
-//   type: "subscription",
-//   method: "Inventory"
-// };
-//
-// // Define a rate limiting rule that matches backorder attempts by non-admin users
-// const addBackorderRule = {
-//   userId: function (userId) {
-//     return Roles.userIsInRole(userId, "createProduct", Reaction.getShopId());
-//   },
-//   type: "method",
-//   method: "inventory/backorder"
-// };
-//
-// // Add the rule, allowing up to 5 messages every 1000 milliseconds.
-// DDPRateLimiter.addRule(addReserveRule, 5, 1000);
-// DDPRateLimiter.addRule(addBackorderRule, 5, 1000);
+export function registerInventory(product) {
+  check(product, Match.OneOf(Schemas.ProductVariant, Schemas.Product));
+  let type;
+  switch (product.type) {
+    case "variant":
+      check(product, Schemas.ProductVariant);
+      type = "variant";
+      break;
+    default:
+      check(product, Schemas.Product);
+      type = "simple";
+  }
+  let totalNewInventory = 0;
+  const productId = type === "variant" ? product.ancestors[0] : product._id;
+  const variants = Catalog.getVariants(productId);
 
-//
-// Inventory methods
-//
+  // we'll check each variant to see if it has been fully registered
+  for (let variant of variants) {
+    let inventory = Inventory.find({
+      productId: productId,
+      variantId: variant._id,
+      shopId: product.shopId
+    });
+    // we'll return this as well
+    let inventoryVariantCount = inventory.count();
+    // if the variant exists already we're remove from the inventoryVariants
+    // so that we don't process it as an insert
+    if (inventoryVariantCount < variant.inventoryQuantity) {
+      let newQty = variant.inventoryQuantity || 0;
+      let i = inventoryVariantCount + 1;
+
+      Logger.info(
+        `inserting ${newQty - inventoryVariantCount
+          } new inventory items for ${variant._id}`
+      );
+
+      const batch = Inventory.
+      _collection.rawCollection().initializeUnorderedBulkOp();
+      while (i <= newQty) {
+        let id = Inventory._makeNewID();
+        batch.insert({
+          _id: id,
+          productId: productId,
+          variantId: variant._id,
+          shopId: product.shopId,
+          createdAt: new Date,
+          updatedAt: new Date,
+          workflow: { // we add this line because `batchInsert` doesn't know
+            status: "new" // about SimpleSchema, so `defaultValue` will not
+          }
+        });
+        i++;
+      }
+
+      // took from: http://guide.meteor.com/collections.html#bulk-data-changes
+      let execute = Meteor.wrapAsync(batch.execute, batch);
+      let inventoryItem = execute();
+      let inserted = inventoryItem.nInserted;
+
+      if (!inserted) { // or maybe `inventory.length === 0`?
+        // throw new Meteor.Error("Inventory Anomaly Detected. Abort! Abort!");
+        return totalNewInventory;
+      }
+      Logger.debug(`registered ${inserted}`);
+      totalNewInventory += inserted;
+    }
+  }
+  // returns the total amount of new inventory created
+  return totalNewInventory;
+}
 
 Meteor.methods({
   /**
-   * inventory/setStatus
-   * @summary sets status from one status to a new status. Defaults to "new" to "reserved"
-   * @param  {Array} cartItems array of objects of type Schemas.CartItems
-   * @param  {String} status optional - sets the inventory workflow status, defaults to "reserved"
-   * @param  {String} currentStatus
-   * @param  {String} notFoundStatus
-   * @todo move this to bulkOp
-   * @return {Number} returns reservationCount
+   * inventory/register
+   * @summary check a product and update Inventory collection with inventory documents.
+   * @param {Object} product - valid Schemas.Product object
+   * @return {Number} - returns the total amount of new inventory created
    */
-  "inventory/setStatus": function (cartItems, status, currentStatus, notFoundStatus) {
-    check(cartItems, [Schemas.CartItem]);
-    check(status, Match.Optional(String));
-    check(currentStatus, Match.Optional(String));
-    check(notFoundStatus, Match.Optional(String));
-    this.unblock();
-
-    // check basic user permissions
-    // if (!Reaction.hasPermission(["guest", "anonymous"])) {
-    //   throw new Meteor.Error(403, "Access Denied");
-    // }
-
-    // set defaults
-    const reservationStatus = status || "reserved"; // change status to options object
-    const defaultStatus = currentStatus || "new"; // default to the "new" status
-    const backorderStatus = notFoundStatus || "backorder"; // change status to options object
-    let reservationCount;
-    Logger.info(`Moving Inventory items from ${defaultStatus} to ${reservationStatus}`);
-
-    // update inventory status for cartItems
-    for (let item of cartItems) {
-      // check of existing reserved inventory for this cart
-      let existingReservations = Inventory.find({
-        productId: item.productId,
-        variantId: item.variants._id,
-        shopId: item.shopId,
-        orderItemId: item._id
-      });
-
-      // define a new reservation
-      let availableInventory = Inventory.find({
-        "productId": item.productId,
-        "variantId": item.variants._id,
-        "shopId": item.shopId,
-        "workflow.status": defaultStatus
-      });
-
-      const totalRequiredQty = item.quantity;
-      const availableInventoryQty = availableInventory.count();
-      let existingReservationQty = existingReservations.count();
-
-      Logger.info("totalRequiredQty", totalRequiredQty);
-      Logger.info("availableInventoryQty", availableInventoryQty);
-
-      // if we don't have existing inventory we create backorders
-      if (totalRequiredQty > availableInventoryQty) {
-        // TODO put in a dashboard setting to allow backorder or altenate handler to be used
-        let backOrderQty = Number(totalRequiredQty - availableInventoryQty - existingReservationQty);
-        Logger.info(`no inventory found, create ${backOrderQty} ${backorderStatus}`);
-        // define a new reservation
-        const reservation = {
-          productId: item.productId,
-          variantId: item.variants._id,
-          shopId: item.shopId,
-          orderItemId: item._id,
-          workflow: {
-            status: backorderStatus
-          }
-        };
-
-        Meteor.call("inventory/backorder", reservation, backOrderQty);
-        existingReservationQty = backOrderQty;
-      }
-      // if we have inventory available, only create additional required reservations
-      Logger.debug("existingReservationQty", existingReservationQty);
-      reservationCount = existingReservationQty;
-      let newReservedQty;
-      if (reservationStatus === "reserved" && defaultStatus === "new") {
-        newReservedQty = totalRequiredQty - existingReservationQty + 1;
-      } else {
-        // when moving from one "reserved" type status, we don't need to deal with existingReservationQty
-        newReservedQty = totalRequiredQty + 1;
-      }
-
-      let i = 1;
-      while (i < newReservedQty) {
-        // updated existing new inventory to be reserved
-        Logger.info(
-          `updating reservation status ${i} of ${newReservedQty - 1}/${totalRequiredQty} items.`);
-        // we should be updating existing inventory here.
-        // backorder process created additional backorder inventory if there
-        // wasn't enough.
-        Inventory.update({
-          "productId": item.productId,
-          "variantId": item.variants._id,
-          "shopId": item.shopId,
-          "workflow.status": defaultStatus
-        }, {
-          $set: {
-            "orderItemId": item._id,
-            "workflow.status": reservationStatus
-          }
-        });
-        reservationCount++;
-        i++;
-      }
+  "inventory/register": function (product) {
+    if (!Reaction.hasPermission("createProduct")) {
+      throw new Meteor.Error(403, "Access Denied");
     }
-    Logger.info(
-      `finished creating ${reservationCount} new ${reservationStatus} reservations`);
-    return reservationCount;
+    registerInventory(product);
   },
   /**
-   * inventory/clearStatus
-   * @summary used to reset status on inventory item (defaults to "new")
-   * @param  {Array} cartItems array of objects Schemas.CartItem
-   * @param  {[type]} status optional reset workflow.status, defaults to "new"
-   * @param  {[type]} currentStatus optional matching workflow.status, defaults to "reserved"
-   * @return {undefined} undefined
-   */
-  "inventory/clearStatus": function (cartItems, status, currentStatus) {
-    check(cartItems, [Schemas.CartItem]);
-    check(status, Match.Optional(String)); // workflow status
-    check(currentStatus, Match.Optional(String));
-    this.unblock();
-
-    // // check basic user permissions
-    // if (!Reaction.hasPermission(["guest", "anonymous"])) {
-    //   throw new Meteor.Error(403, "Access Denied");
-    // }
-
-    // optional workflow status or default to "new"
-    let newStatus = status || "new";
-    let oldStatus = currentStatus || "reserved";
-
-    // remove each cart item in inventory
-    for (let item of cartItems) {
-      // check of existing reserved inventory for this cart
-      let existingReservations = Inventory.find({
-        "productId": item.productId,
-        "variantId": item.variants._id,
-        "shopId": item.shopId,
-        "orderItemId": item._id,
-        "workflow.status": oldStatus
-      });
-      let i = existingReservations.count();
-      // reset existing cartItem reservations
-      while (i <= item.quantity) {
-        Inventory.update({
-          "productId": item.productId,
-          "variantId": item.variants._id,
-          "shopId": item.shopId,
-          "orderItemId": item._id,
-          "workflow.status": oldStatus
-        }, {
-          $set: {
-            "orderItemId": "", // clear order/cart
-            "workflow.status": newStatus // reset status
-          }
-        });
-        i++;
-      }
-    }
-    Logger.info("inventory/clearReserve", newStatus);
-  },
-  /**
-   * inventory/clearReserve
-   * @summary resets "reserved" items to "new"
-   * @param  {Array} cartItems array of objects Schemas.CartItem
-   * @return {undefined}
-   */
-  "inventory/clearReserve": function (cartItems) {
-    check(cartItems, [Schemas.CartItem]);
-    return Meteor.call("inventory/clearStatus", cartItems);
-  },
-  /**
-   * inventory/clearReserve
-   * converts new items to reserved, or backorders
-   * @param  {Array} cartItems array of objects Schemas.CartItem
-   * @return {undefined}
-   */
-  "inventory/addReserve": function (cartItems) {
-    check(cartItems, [Schemas.CartItem]);
-    return Meteor.call("inventory/setStatus", cartItems);
-  },
-  /**
-   * inventory/backorder
-   * @summary is used by the cart process to create a new Inventory
-   * backorder item, but this could be used for inserting any
-   * custom inventory.
+   * inventory/adjust
+   * @summary adjust existing inventory when changes are made we get the
+   * inventoryQuantity for each product variant, and compare the qty to the qty
+   * in the inventory records we will add inventoryItems as needed to have the
+   * same amount as the inventoryQuantity but when deleting, we'll refuse to
+   * delete anything not workflow.status = "new"
    *
-   * A note on DDP Limits.
-   * As these are wide open we defined some ddp limiting rules http://docs.meteor.com/#/full/ddpratelimiter
-   *
-   * @param {Object} reservation Schemas.Inventory
-   * @param {Number} backOrderQty number of backorder items to create
-   * @returns {Number} number of inserted backorder documents
+   * @param  {Object} product - Schemas.Product object
+   * @return {[undefined]} returns undefined
    */
-  "inventory/backorder": function (reservation, backOrderQty) {
-    check(reservation, Schemas.Inventory);
-    check(backOrderQty, Number);
-    this.unblock();
-
-    // this use case could happen then mergeCart is fires. We don't add anything
-    // or remove, just item owner changed. We need to add this check here
-    // because of bulk operation. It thows exception if nothing to operate.
-    if (backOrderQty === 0) {
-      return 0;
+  "inventory/adjust": function (product) { // TODO: this should be variant
+    check(product, Match.OneOf(Schemas.Product, Schemas.ProductVariant));
+    let type;
+    let results;
+    // adds or updates inventory collection with this product
+    switch (product.type) {
+      case "variant":
+        check(product, Schemas.ProductVariant);
+        type = "variant";
+        break;
+      default:
+        check(product, Schemas.Product);
+        type = "simple";
     }
+    // user needs createProduct permission to adjust inventory
+    if (!Reaction.hasPermission("createProduct")) {
+      throw new Meteor.Error(403, "Access Denied");
+    }
+    // this.unblock();
 
-    // TODO: need to look carefully and understand is it possible ho have a
-    // negative `backOrderQty` value here?
-
-    // check basic user permissions
-    // if (!Reaction.hasPermission(["guest","anonymous"])) {
-    //   throw new Meteor.Error(403, "Access Denied");
-    // }
-
-    // set defaults
-    let newReservation = reservation;
-    if (!newReservation.workflow) {
-      newReservation.workflow = {
-        status: "backorder"
+    // Quantity and variants of this product's variant inventory
+    if (type === "variant") {
+      const variant = {
+        _id: product._id,
+        qty: product.inventoryQuantity || 0
       };
+
+      const inventory = Inventory.find({
+        productId: product.ancestors[0],
+        variantId: product._id
+      });
+      const itemCount = inventory.count();
+
+      if (itemCount !== variant.qty) {
+        if (itemCount < variant.qty) {
+          // we need to register some new variants to inventory
+          results = itemCount + Meteor.call("inventory/register", product);
+        } else if (itemCount > variant.qty) {
+          // determine how many records to delete
+          const removeQty = itemCount - variant.qty;
+          // we're only going to delete records that are new
+          const removeInventory = Inventory.find({
+            "variantId": variant._id,
+            "workflow.status": "new"
+          }, {
+            sort: {
+              updatedAt: -1
+            },
+            limit: removeQty
+          }).fetch();
+
+          results = itemCount;
+          // delete latest inventory "status:new" records
+          for (let inventoryItem of removeInventory) {
+            results -= Meteor.call("inventory/remove", inventoryItem);
+            // we could add handling for the case when aren't enough "new" items
+          }
+        }
+        Logger.info(
+          `adjust variant ${variant._id} from ${itemCount} to ${results}`
+        );
+      }
     }
-
-    // insert backorder
-    let i = 0;
-    const batch = Inventory.
-      _collection.rawCollection().initializeUnorderedBulkOp();
-    while (i < backOrderQty) {
-      let id = Inventory._makeNewID();
-      batch.insert(Object.assign({ _id: id }, newReservation));
-      i++;
-    }
-
-    const execute = Meteor.wrapAsync(batch.execute, batch);
-    const inventoryBackorder = execute();
-    const inserted = inventoryBackorder.nInserted;
-    Logger.info(
-      `created ${inserted} backorder records for product ${
-        newReservation.productId}, variant ${newReservation.variantId}`);
-
-    return inserted;
-  },
-  //
-  // send low stock warnings
-  //
-  "inventory/lowStock": function (product) {
-    check(product, Schemas.Product);
-    // WIP placeholder
-    Logger.info("inventory/lowStock");
   }
+
 });
